refsw: sync up to 01.06.2018.
refsw: 8c4730f4da5623bd1c7abdf29db5b88eb0d1eb79

+++ BSEAV/linux/driver/usermode/bcm_driver.h
@@ -121,7 +121,7 @@ typedef struct t_bcm_linux_mmap_access
 {
     unsigned pid; /* process id of client */
     struct {
-        unsigned offset; /* set all to zero to revoke access */
+        uint64_t offset; /* set all to zero to revoke access */
         unsigned size;
     } heap[BCM_MAX_HEAPS];
 } t_bcm_linux_mmap_access;
